handle token refresh in fetcher
